fix msvc warnings/errors

+++ nucleus/tile_scheduler/Scheduler.cpp
@@ -240,14 +240,14 @@ void Scheduler::persist_tiles()
 
 void Scheduler::schedule_update()
 {
-    assert(m_update_timeout < std::numeric_limits<int>::max());
+    assert(m_update_timeout < unsigned(std::numeric_limits<int>::max()));
     if (m_enabled && !m_update_timer->isActive())
         m_update_timer->start(int(m_update_timeout));
 }
 
 void Scheduler::schedule_purge()
 {
-    assert(m_purge_timeout < std::numeric_limits<int>::max());
+    assert(m_purge_timeout < unsigned(std::numeric_limits<int>::max()));
     if (m_enabled && !m_purge_timer->isActive()) {
         m_purge_timer->start(int(m_purge_timeout));
     }
@@ -255,7 +255,7 @@ void Scheduler::schedule_purge()
 
 void Scheduler::schedule_persist()
 {
-    assert(m_persist_timeout < std::numeric_limits<int>::max());
+    assert(m_persist_timeout < unsigned(std::numeric_limits<int>::max()));
     if (!m_persist_timer->isActive()) {
         m_persist_timer->start(int(m_persist_timeout));
     }
@@ -309,7 +309,7 @@ unsigned int Scheduler::persist_timeout() const
 
 void Scheduler::set_persist_timeout(unsigned int new_persist_timeout)
 {
-    assert(new_persist_timeout < std::numeric_limits<int>::max());
+    assert(new_persist_timeout < unsigned(std::numeric_limits<int>::max()));
     m_persist_timeout = new_persist_timeout;
 
     if (m_persist_timer->isActive()) {
@@ -358,7 +358,7 @@ std::filesystem::path Scheduler::disk_cache_path()
 
 void Scheduler::set_purge_timeout(unsigned int new_purge_timeout)
 {
-    assert(new_purge_timeout < std::numeric_limits<int>::max());
+    assert(new_purge_timeout < unsigned(std::numeric_limits<int>::max()));
     m_purge_timeout = new_purge_timeout;
 
     if (m_purge_timer->isActive()) {
@@ -399,7 +399,7 @@ void Scheduler::set_enabled(bool new_enabled)
 
 void Scheduler::set_update_timeout(unsigned new_update_timeout)
 {
-    assert(m_update_timeout < std::numeric_limits<int>::max());
+    assert(m_update_timeout < unsigned(std::numeric_limits<int>::max()));
     m_update_timeout = new_update_timeout;
     if (m_update_timer->isActive()) {
         m_update_timer->start(m_update_timeout);
